# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.16.0

package db

import (
	"database/sql"

	"github.com/google/uuid"
)

type User struct {
	ID          uuid.UUID
	Email       string
	Password    string
	Username    string
	CreatedAt   sql.NullTime
	Avatar      sql.NullString
	Description sql.NullString
	Followers   int32
}

type Video struct {
	ID          uuid.UUID
	Path        string
	UploadedAt  sql.NullTime
	OwnerID     uuid.UUID
	Thumbnail   string
	Title       string
	Description string
}