# internal/rest/openapi3.yaml
openapi: 3.0.0
components:
  schemas:
    User:
      type: object
      properties:
        avatar:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        description:
          type: string
          nullable: true
        email:
          type: string
        followers:
          type: integer
        id:
          type: string
          format: uuid
    Video:
      type: object
      properties:
        description:
          type: string
        id:
          type: string
          format: uuid
        owner:
          $ref: '#/components/schemas/User'
        path:
          type: string
        thumbnail:
          type: string
        title:
          type: string
        uploaded_at:
          type: string
          format: date-time
          nullable: true
  requestBodies:
    AuthRequest:
      description: Request used for creating an account.
      required: true
      content:
        application/json:
          schema:
            properties:
              email:
                type: string
                minLength: 1
              password:
                type: string
                minLength: 1
    RenewTokenRequest:
      description: Request used for renewing jwt token.
      required: true
      content:
        application/json:
          schema:
            properties:
              token:
                type: string
                minLength: 1
    SearchVideoRequest:
      description: Request used for searching a video.
      required: true
      content:
        application/json:
          schema:
            properties:
              from:
                type: integer
                format: int64
                default: 0
              size:
                type: integer
                format: int64
                default: 10
              title:
                type: string
                minLength: 1
  responses:
    ErrorResponse:
      description: Response when errors happen.
      content:
        application/json:
          schema:
            properties:
              error:
                type: string
    SearchVideoResponse:
      description: Response returned back after searching for any videos.
      content:
        application/json:
          schema:
            properties:
              tasks:
                type: array
                items:
                  $ref: '#/components/schemas/Video'
              total:
                type: integer
                format: int64
    UserResponse:
      description: Login response.
      content:
        application/json:
          schema:
            properties:
              access_token:
                type: string
              refresh_token:
                type: string
              user:
                $ref: '#/components/schemas/User'
info:
  title: Ntube
  description: Ntube REST APIs
  contact:
    url: https://github.com/nei7/ntube
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 0.0.1
paths:
  /login:
    post:
      operationId: Login
      requestBody:
        $ref: '#/components/requestBodies/AuthRequest'
      responses:
        "200":
          $ref: '#/components/responses/UserResponse'
        "400":
          $ref: '#/components/responses/ErrorResponse'
        "500":
          $ref: '#/components/responses/ErrorResponse'
  /signup:
    post:
      operationId: Signup
      requestBody:
        $ref: '#/components/requestBodies/AuthRequest'
      responses:
        "201":
          $ref: '#/components/responses/UserResponse'
        "400":
          $ref: '#/components/responses/ErrorResponse'
        "500":
          $ref: '#/components/responses/ErrorResponse'
  /videos/search:
    post:
      operationId: SearchVideos
      requestBody:
        $ref: '#/components/requestBodies/SearchVideoRequest'
      responses:
        "200":
          $ref: '#/components/responses/SearchVideoResponse'
        "400":
          $ref: '#/components/responses/ErrorResponse'
        "500":
          $ref: '#/components/responses/ErrorResponse'
servers:
- url: http://127.0.0.1:3001
  description: Rest server